refactored rabbitmq codebase
