+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-rabbitmq-e2e/src/test/java/org/apache/seatunnel/e2e/connector/rabbitmq/RabbitmqIT.java
@@ -28,7 +28,7 @@ import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.api.table.type.SeaTunnelRowType;
 import org.apache.seatunnel.common.Handover;
 import org.apache.seatunnel.connectors.seatunnel.rabbitmq.client.RabbitmqClient;
-import org.apache.seatunnel.connectors.seatunnel.rabbitmq.config.RabbitmqConfig;
+import org.apache.seatunnel.connectors.seatunnel.rabbitmq.config.RabbitmqBaseOptions;
 import org.apache.seatunnel.e2e.common.TestResource;
 import org.apache.seatunnel.e2e.common.TestSuiteBase;
 import org.apache.seatunnel.e2e.common.container.TestContainer;
@@ -181,7 +181,7 @@ public class RabbitmqIT extends TestSuiteBase implements TestResource {
 
     private void initRabbitMQ() {
         try {
-            RabbitmqConfig config = new RabbitmqConfig();
+            RabbitmqBaseOptions config = new RabbitmqBaseOptions();
             config.setHost(rabbitmqContainer.getHost());
             config.setPort(rabbitmqContainer.getFirstMappedPort());
             config.setQueueName(QUEUE_NAME);
@@ -200,7 +200,7 @@ public class RabbitmqIT extends TestSuiteBase implements TestResource {
     private RabbitmqClient initSinkRabbitMQ() {
 
         try {
-            RabbitmqConfig config = new RabbitmqConfig();
+            RabbitmqBaseOptions config = new RabbitmqBaseOptions();
             config.setHost(rabbitmqContainer.getHost());
             config.setPort(rabbitmqContainer.getFirstMappedPort());
             config.setQueueName(SINK_QUEUE_NAME);
